Get Tasks (#1)
* Create executable for our gem

We want our gem to be executed from the command line so we add a script
in the bin folder to run the main class, passing in the arguments.

* Create fixture for tasks

We create a fixture module for tasks since we'll need to test json
parsing for the get tasks command. We also specify that we don't want to
flag big methods inside our specs to the linter since we plan to have
many such fixtures.

* Parse/format tasks json array

We implement the basic parsing and formatting of tasks for our `get`
command.

* Add webmock to gems

We plan to uses that gem to mock the cli's api calls during tests.

* Implement get task requests

We add the missing get tasks request and mock the requests instead of
the whole client object.

* Configure api secrets

We inject api secrets as environment variables for now. We plan to
create a configuration file / commands in the future.